requirements.txt
appdirs
click
Flask
itsdangerous
Jinja2
MarkupSafe
packaging
pyparsing
six
Werkzeug
flask_sqlalchemy
Flask-Migrate
configparser
deepmerge
pathlib
PyYAML
psycopg2